Unify claim utilities in `raiden.utils.claim` package
This unifies the code from
- `raiden.claim`
- `tools.raiddit.generate_claims`
in `raiden.utils.claims`.

`tools/raiddit/generate_claims.py` now only contains the CLI part.

This is useful since the `tools` package is not part of the Raiden source
distribution and therefore code from it can't be used outside a source
checkout (e.g. scenario player).
