+++ raiden/storage/serialization/types.py
@@ -84,7 +84,7 @@ from raiden.utils.typing import (
 
 MESSAGE_NAME_TO_QUALIFIED_NAME = {
     "AuthenticatedMessage": "raiden.messages.abstract.AuthenticatedMessage",
-    "Claim": "raiden.claim.Claim",
+    "Claim": "raiden.transfer.state.Claim",
     "Delivered": "raiden.messages.synchronization.Delivered",
     "EnvelopeMessage": "raiden.messages.transfers.EnvelopeMessage",
     "LockedTransferBase": "raiden.messages.transfers.LockedTransferBase",
